requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==24.2.0
black==24.10.0
certifi==2025.4.26
click==8.1.7
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.83.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.11.5
pydantic-core==2.33.2
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
yarl==1.18.3
